Fix Swagger documentation
Commenting out bearerAuth is enough to tell me that you want to replace
it with something else, but definitely not enough to make sure you don't
crash the Swagger docs (and beyond). I went through the entire app, and
switched the 10 or so remaining endpoints to use the new cookieAuth.
I had to fill in missing schemas for the new endpoints, and update the
documentation to match reality. I shortened the getUserAvatarSchema;
it was missing type and was unnecessarily specific. Fastify takes
care of a lot of stuff by default, no need to be so specific. All
images are binary data. Easier, and shorter. I have also fixed
a styling error in the Swagger docs. No biggie, but it was annoying.