Fix keyboard event listeners not releasing
There was a tiny bug in Pong Tournament, and consequently copied over to
Tetris Tournament, where keyboard event listeners were not released
when leaving the tournament view. This caused issues when navigating
back to the tournament view, as multiple listeners would be active
at the same time, leading to unexpected behavior. Mostly the keys could
not be used to input text into the `room name` field. This is now fixed
in both tournaments, and the keyboard event listeners are being properly
released after leaving the tournament view.