Add Avalanche Fuji local network setup
Using avalanche-cli this is the simplest setup I could come up with.
Every other solution I came across is either so generic, it cannot be
reasonably argued it has anything to do the the Avalanche network, or
requires AVAX testnet tokens, which I've tried to get. For days. Nothing
more than a total waste of time. This runs locally, it should a lot faster
than the actual online variant, and it is pre-funded with test accounts.
No need to go looking for AVAX tokens.