Save tournaments data to local blockchain
In effect, the blockchain data persists in the volume ONLY if a snapshot
of the blockchain is saved before stopping the node. avalanche-cli other
starts a new blockchain at a new address even the the old data still
exists in the volume. Snapshot cannot be save while the node is running.
To save a snapshot, use `make blockchain-stop`.