Do a security check, fix unsafe DOM methods
If and when you find the time, take a look at the Security Audit doc.
It's not an uninteresting read, if I say so myself. After fixing some
issues around unsafe usage of innerHTML, we haven't come out of it even
too bad, if you look at the results. There is always something to improve,
but at this point in time, we have definitely secured our app more than
enough for how it's going to be used. If this ever went into production,
we would naturally have to limit or even disable the Swagger docs, limit
how much we disclose about the app's architecture in our error messages,
and so on. But those are mostly details, that we don't necessarily have
to deal with right now. We SHOULD though be aware of how and why we could
do a little better in PROD.