Remove redundant/internal data from tournament JSON
The IDs are a trifle really, but there is no reason to reveal them
publically when it is not needed. Internal DB data should stay internal.
Revealing too much of the app architecture is not a safe strategy.
And since all tournament players must be registered users, there is no need
to store player_is_user flags in tournament games.