Fix switching to another port for backend and nginx
The CORS was working exactly as it should--preventing fetching data from
port 443 while the application was running on port 1488. Potential
security risk. The problem was there was a lot of places in the codebase
(such as at the /stats route) that was fetching data using just the localhost
without any port (...localhost/.....). The the default port 443 was supplied,
and CORS interfered a stopped the transfer. I've added an ENV variable
to set the port automatically in the frontend in the future. To change
the port in the future:
1. Change the NGINX_HTTPS_PORT variable in the .env file.
2. Change the port mapping in the docker-compose.yaml file.
3. Change the port in the NGINX server.conf file if necessary.