Add flag file to Makefile for cert generation
The certs were not being generated. I didn't much investigate why,
I went the easy route and just added a empty flag file to the Makefile to
force the generation of the certs. If you want to recreate the certs,
just delete the whole `/secrets` directory, run the `check_secrets.sh`
script you can find in our chatroom on Slack. and make again.