Update chat behavior during gameplay
For all games (including the tournament ones) now:
1. Chat panel automatically hides at the start of the game.
2. Chat is otherwise enabled and accessible during gameplay.
3. Games section in Chat is now hidden during gameplay. This prevents the
potential mess of leaving the current game to start a new game somewhere
else (and all the slew of unhandled states it entails). The logic can
obviously be written, and the new states handled, but I've just about had
enough of this project for the time being. So for now, the easiest solution
is to hide the Games section during gameplay. Hence, the potential mess
can never happen.