Disable user input during tournament countdowns
This is the only effective way I could think of to prevent unexpected
behavior when users try to interact with the game during critical phases
like countdowns before matches start. By disabling the interactions, we ensure
that users cannot inadvertently interfere with the game state. I mean, if
you really really tried, you could still cause havoc, but at the very least
it now much harder for a user to leave a room during countdown. A safer
solution would be to remove the countdowns entirely, of course. But during the
tournaments, the countdown is quite necessary to let next players prepare,
and be ready to start playing.